- Fixed build without opencv and with opencv 1.0.0 - Fixed warning in debian/copyright - Fixed annoying warnings in lib3ds that were there since 1917. - Copied fix in RealSchur for the weird GCC warning.

+++ libs/opengl/src/lib3ds/chunk.c
@@ -62,7 +62,7 @@ lib3ds_chunk_debug_dump(Lib3dsChunk *c)
       lib3ds_chunk_level,
       lib3ds_chunk_name(c->chunk),
       c->chunk,
-      c->size
+      (long unsigned int)c->size
     );
   }
 }
